% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toastifyExamples.R
\name{shinyToastifyExample}
\alias{shinyToastifyExample}
\title{Run a Shiny Toastify example}
\usage{
shinyToastifyExample(example, display.mode = "showcase", ...)
}
\arguments{
\item{example}{example name}

\item{display.mode}{the display mode to use when running the example; see
\code{\link[shiny:runApp]{runApp}}}

\item{...}{arguments passed to \code{\link[shiny:runApp]{runApp}}}
}
\value{
No return value, just launches a Shiny app.
}
\description{
A function to run examples of Shiny apps using the
  \code{shinyToastify} package.
}
\examples{
if(interactive()){
  shinyToastifyExample("bodyClassName")
}
if(interactive()){
  shinyToastifyExample("className")
}
if(interactive()){
  shinyToastifyExample("progressClassName")
}
if(interactive()){
  shinyToastifyExample("TypeAndTransition")
}
}
