% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toastify.R
\name{useShinyToastify}
\alias{useShinyToastify}
\title{Use Shiny toastify}
\usage{
useShinyToastify()
}
\value{
An object of class \code{shiny.tag.list}.
}
\description{
This function must be called once in your Shiny \code{ui} to
  allow to use \code{\link{showToast}}.
}
