% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toastifyExamples.R
\name{shinyToastifyExamples}
\alias{shinyToastifyExamples}
\title{Shiny Toastify examples}
\usage{
shinyToastifyExamples()
}
\value{
No return value, just prints a message listing the example names.
}
\description{
List of examples.
}
\examples{
shinyToastifyExamples()
if(interactive()){
  shinyToastifyExample("TypeAndTransition")
}
}
